% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{calSens}
\alias{calSens}
\title{Calculate sensitivity, specificity}
\usage{
calSens(x, y)
}
\arguments{
\item{x}{A numeric vector as a predictor}

\item{y}{A numeric vector as a result}
}
\value{
A data.frame
}
\description{
Calculate sensitivity, specificity
}
\examples{
require(moonBook)
calSens(radial$height,radial$male)
}
