% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{makeCoord}
\alias{makeCoord}
\title{Calculate x, y coordinate for ROC curve}
\usage{
makeCoord(x, no = 1)
}
\arguments{
\item{x}{An object of class multipleROC}

\item{no}{Integer}
}
\value{
A data.frame
}
\description{
Calculate x, y coordinate for ROC curve
}
\examples{
x=multipleROC(form=am~wt,data=mtcars)
makeCoord(x)
}
