% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{makeLabels}
\alias{makeLabels}
\title{Make labels for ROC curve}
\usage{
makeLabels(x, no = 1)
}
\arguments{
\item{x}{An object of class multipleROC}

\item{no}{Integer}
}
\value{
A data.frame
}
\description{
Make labels for ROC curve
}
\examples{
x=multipleROC(am~wt,data=mtcars)
makeLabels(x)
}
