% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePPTxList_ROC.R
\name{makePPTxList_ROC}
\alias{makePPTxList_ROC}
\title{Make Powerpoint List for ROC curve Analysis}
\usage{
makePPTxList_ROC(
  yvar = "male",
  xvars = c("weight", "height", "age"),
  dataname = "radial",
  multiple = FALSE,
  separate = TRUE,
  vanilla = TRUE,
  show.line = FALSE,
  show.points = TRUE,
  show.eta = TRUE,
  show.sens = TRUE,
  show.AUC = TRUE
)
}
\arguments{
\item{yvar}{Name of dependent variable}

\item{xvars}{Names of independent variables}

\item{dataname}{Name of data}

\item{multiple}{logical Whether or not analyze as multiple predictor model}

\item{separate}{logical Whether or not plot independent variables separately}

\item{vanilla}{logical Whether or not make vanilla table}

\item{show.line, show.points, show.eta, show.sens, show.AUC}{logical Arguments to be passed to plot_ROC()}
}
\description{
Make Powerpoint List for ROC curve Analysis
}
\examples{
require(moonBook)
require(ggplot2)
require(multipleROC)
result=makePPTxList_ROC()
result=makePPTxList_ROC(multiple=TRUE)
}
