% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{multipleROC}
\alias{multipleROC}
\alias{plot.multipleROC}
\title{Do a ROC Curve Analysis}
\usage{
multipleROC(formula, data, plot = TRUE)

\method{plot}{multipleROC}(x, ...)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data.frame}

\item{plot}{logical Whether or not draw plot}

\item{x}{\code{multipleROC}}

\item{...}{Further arguments to be passed to plot_ROC}
}
\value{
A list of class of multipleROC with elements
\item{fit}{A class of glm}
\item{df}{a data.frame}
\item{cutpoint}{Numeric. cutpoint}
\item{sens}{label with sensitivity, specificity}
\item{auc}{numeric. area under curve value}
\item{cutoff}{A data.fame with best cutoff value(s)}
}
\description{
Do a ROC Curve Analysis
}
\examples{
multipleROC(am~wt,data=mtcars)
}
