% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{multipleROC2roc}
\alias{multipleROC2roc}
\title{Convert A list from ROC() to roc object}
\usage{
multipleROC2roc(x)
}
\arguments{
\item{x}{An object of class multipleROC}
}
\description{
Convert A list from ROC() to roc object
}
