% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{plot_ROC}
\alias{plot_ROC}
\title{Draw ROC curves}
\usage{
plot_ROC(
  x,
  show.points = TRUE,
  show.eta = TRUE,
  show.sens = TRUE,
  show.AUC = TRUE,
  facet = FALSE
)
}
\arguments{
\item{x}{A list of class multipleROC}

\item{show.points}{logical}

\item{show.eta}{logical}

\item{show.sens}{logical}

\item{show.AUC}{logical}

\item{facet}{logical}
}
\value{
A ggplot
}
\description{
Draw ROC curves
}
\examples{
require(moonBook)
a=multipleROC(male~height,data=radial)
plot(a)
b=multipleROC(male~age,data=radial)
plot_ROC(list(a,b))
c=multipleROC(form=male~weight,data=radial)
plot_ROC(list(a,b,c),show.eta=FALSE,show.sens=FALSE)
plot_ROC(list(a,b,c),facet=TRUE)
require(ggplot2)
plot_ROC(list(a,b,c))+facet_grid(no~.)
}
