% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{plot_ROC2}
\alias{plot_ROC2}
\title{Draw ROC curve with variable names}
\usage{
plot_ROC2(yvar, xvars, dataname, ...)
}
\arguments{
\item{yvar}{Name of dependent variable}

\item{xvars}{Names of independent variables}

\item{dataname}{Name of data}

\item{...}{Further arguments to be passed to plot_ROC}
}
\description{
Draw ROC curve with variable names
}
\examples{
require(moonBook)
plot_ROC2(yvar="male",xvars=c("weight","height","age"),dataname="radial")
}
