% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{simpleAUC}
\alias{simpleAUC}
\title{Calculate area under curve}
\usage{
simpleAUC(df)
}
\arguments{
\item{df}{A data.frame Result of calSens()}
}
\value{
A numeric of area under curve
}
\description{
Calculate area under curve
}
