% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleRoc.R
\name{step_ROC}
\alias{step_ROC}
\title{Perform multiple logistic regression with stepwise}
\usage{
step_ROC(formula, data, plot = TRUE, trace = 0, ...)
}
\arguments{
\item{formula}{A formula for logistic regression}

\item{data}{A data.frame}

\item{plot}{logical If true, return a ggplot}

\item{trace}{if positive, information is printed during the running of step. Larger values may give more detailed information.}

\item{...}{Further arguments to be passed to plot_ROC()}
}
\value{
A ggplot or an object of class anova
}
\description{
Perform multiple logistic regression with stepwise
}
\examples{
require(moonBook)
step_ROC(male~weight+height+age,data=radial)
step_ROC(male~weight+height+age,data=radial,plot=FALSE)
}
